+++ src/modules/actor/sheet/cpr-demon-sheet.js
@@ -1,10 +1,10 @@
 import CPRChat from "../../chat/cpr-chat.js";
 import SystemUtils from "../../utils/cpr-systemUtils.js";
 import createImageContextMenu from "../../utils/cpr-imageContextMenu.js";
+import enrichDocumentHTML from "../../utils/cpr-enrich-html.js";
 
 const { HandlebarsApplicationMixin } = foundry.applications.api;
 const { ActorSheetV2 } = foundry.applications.sheets;
-const TextEditor = foundry.applications.ux.TextEditor.implementation;
 
 /**
  * Implement the Demon sheet, which extends ActorSheet directly from Foundry. This does
@@ -68,9 +68,9 @@ export default class CPRDemonActorSheet extends HandlebarsApplicationMixin(
   async _prepareContext(options) {
     const sheetData = await super._prepareContext(options);
     sheetData.enrichedHTML = sheetData.enrichedHTML ?? {};
-    sheetData.enrichedHTML.notes = await TextEditor.enrichHTML(
+    sheetData.enrichedHTML.notes = await enrichDocumentHTML(
       this.actor.system.notes,
-      { async: true }
+      this.actor
     );
     return sheetData;
   }